+++ oefdb/schema_validation/schema.py
@@ -1,76 +1,13 @@
 from __future__ import annotations
 
-import pprint
-import typing
-from typing import List, Optional
-
 import toml
-from pydantic import BaseModel, Field, validator
 
-from oefdb.schema_validation.cell_validators import ALL_VALIDATORS
+from oefdb.schema_validation.column_schema import ColumnSchema
 from oefdb.validators._typing import validator_result_type
 
 
-class CellValidator(BaseModel):
-    validator_name: str
-    validator_function: typing.Callable[[str], validator_result_type]
-
-    def validate(self, obj) -> validator_result_type:
-        return self.validator_function(obj)
-
-
-class ColumnConfiguration(BaseModel):
-    name: str
-    validator_strings: list[str] = Field(alias="validators")
-    allow_empty: bool
-    legal_values: str | None = None
-
-    def validators(self) -> list[CellValidator]:
-        return [
-            CellValidator(
-                validator_name=validator_name,
-                validator_function=ALL_VALIDATORS[validator_name],
-            )
-            for validator_name in self.validator_strings
-        ]
-
-    @validator("validator_strings")
-    def check_validator_is_defined(cls, v):
-        for validator in v:
-            if validator not in ALL_VALIDATORS:
-                keys = list(ALL_VALIDATORS.keys())
-                raise ValueError(
-                    f"Wrong validator '{validator}' provided. Validator must one of: {keys}"
-                )
-        return v
-
-    """
-    Returns
-    """
-
-    def validate_cell(self, cell) -> (bool, dict):
-        if cell == "" and self.allow_empty:
-            print("Accepting empty cell")
-            return True, []
-
-        # todo better errors
-        all_errors = {}
-
-        for validator in self.validators():
-            valid, errors = validator.validate(cell)
-            if valid:
-                pass
-            else:
-                all_errors[f"{validator.validator_name}"] = errors
-
-        pprint.pp(all_errors)
-        if all_errors:
-            return False, all_errors
-        return True, all_errors
-
-
-class ColumnSchema(BaseModel):
-    columns: list[ColumnConfiguration]
+class Schema(BaseModel):
+    columns: list[ColumnSchema]
 
     def validate_single_row(self, row):
         all_errors = {}
@@ -141,12 +78,12 @@ class ColumnSchema(BaseModel):
         return self.validate_rows(rows)
 
     @staticmethod
-    def load_schema_definition(file_path: str) -> ColumnSchema:
+    def load_schema_definition(file_path: str) -> Schema:
         with open(file_path) as f:
             toml_fle = f.read()
 
         configuration = toml.loads(toml_fle)
 
-        columns = [ColumnConfiguration(**conf) for conf in configuration["columns"]]
+        columns = [ColumnSchema(**conf) for conf in configuration["columns"]]
 
-        return ColumnSchema(columns=columns)
+        return Schema(columns=columns)
