+++ pyproject.toml
@@ -32,6 +32,7 @@ pyperclip = ">=1.8.2,<2.0.0"
 python-dotenv = ">=0.19.2,<1.0.0"
 click = ">=8.0.3,<9.0.0"
 mypy = ">=0.940,<1.0.0"
+pydantic = "^1.9.1"
 
 [tool.poetry.dev-dependencies]
 coverage = "*"
@@ -46,6 +47,8 @@ jupyter = "^1.0.0"
 ipykernel = "^6.6.0"
 jupytext = "^1.13.3"
 types-click = "^7.1.8"
+types-toml = "^0.10.7"
+
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
